+++ prediction_service/views.py
@@ -1,7 +1,7 @@
 from .modules import api
 from .modules import utils
 from django.http import HttpResponse, JsonResponse
-from enrollment_predictions.enrollment_predictions import enrollment_predictions
+from enrollment_predictions.enrollment_predictions import enrollment_predictions, most_recent_enrollments
 import json
 import pandas as pd
 
@@ -75,33 +75,6 @@ def predict(request):
         # score = perform_decision_tree()
     #TODO: Return predictions to backend (json)
     '''
-
-def most_recent_enrollments(historic_schedules, courses):
-    historic_schedules = pd.DataFrame(historic_schedules)
-    historic_schedules['Course'] = historic_schedules['Subj'] + historic_schedules["Num"]
-    term_month = historic_schedules['Term'].astype(int) % 100
-    season_mapping = {5: 1, 9: 2, 1: 3}
-    historic_schedules['Season'] = term_month.map(season_mapping)
-
-    courses = pd.DataFrame(courses)
-    courses['Course'] = courses['Subj'] + courses["Num"]
-    term_month = courses['Term'].astype(int) % 100
-    season_mapping = {5: 1, 9: 2, 1: 3}
-    courses['Season'] = term_month.map(season_mapping)
-
-    predictions = []
-    for ind in courses.index:
-        past_offerings = historic_schedules[historic_schedules["Course"] == courses['Course'][ind]]
-        past_offerings = past_offerings[past_offerings["Season"] == courses['Season'][ind]]
-        most_recent_offering = past_offerings[past_offerings["Term"] == past_offerings["Term"].max()].iloc[0]
-
-        prediction = {
-            "course" : courses['Course'][ind],
-            "estimate" : most_recent_offering["Enrolled"]
-        }
-        predictions.append(prediction)
-
-    return(predictions)
 
 if __name__ == '__main__':
     with open('data/client_data/schedules.json', 'r', encoding='utf-8') as fh:
